+++ lib/batch_processor/collection.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+# A `Collection` takes input to validate and build a (possibly ordered) list of items to process with the Batch's job.
 module BatchProcessor
   class Collection < Spicerack::InputModel
     def items