+++ spec/integration/default_batch_integration_spec.rb
@@ -2,10 +2,10 @@
 
 RSpec.describe DefaultBatch, type: :integration do
   context "without any arguments" do
-    subject(:process) { described_class.process }
+    subject(:process!) { described_class.process! }
 
     it "raises" do
-      expect { process }.to raise_error BatchProcessor::CollectionEmptyError
+      expect { process! }.to raise_error BatchProcessor::CollectionEmptyError
     end
   end
 end